Expose AWS object as same as upstream aws-sdk-js do so.
Factory methods/type aliases are added as extension methods.
